// src/components/exhibition/dto/dashboard-viewer-of-exhibitions.dto.ts
/**
 * COMPAON API
 * API definition for COMPAON
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: compaon@dalavina.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface DashboardViewerOfExhibitions {
    exhibition_name: string;
    view: number;
}